Move out BoxData and rename into Input
